Provide a bower package file and set at 0.4.0
This update provides Bower support with a package file, and fixes #9.

Additionally, the openfb.js is moved into a distribution directory,
providing a layout where a minified version could be provided if
desired.

You will also likely want to create and push a git tag for the version
number, so bower can pick it up. See:
https://github.com/bower/bower.json-spec#version

To register the component so that it is discoverable, use the following:

    $ bower register openfb https://github.com/ccoenraets/OpenFB